refactor(identity-core): ActorContextHolder 자동 구성 책임 이동
