+++ bootstrap/resource-application-starter/src/main/java/com/seatliberator/seatliberator/bootstrap/autoconfigure/ResourceServerAuthorizeAutoConfiguration.java
@@ -1,16 +1,10 @@
 package com.seatliberator.seatliberator.bootstrap.autoconfigure;
 
 import com.seatliberator.seatliberator.bootstrap.security.ActorContextBindingFilter;
-import com.seatliberator.seatliberator.bootstrap.security.ActorContextJwtAuthenticationConverter;
-import com.seatliberator.seatliberator.bootstrap.security.JwtAuthenticationTokenConverter;
 import com.seatliberator.seatliberator.bootstrap.security.customizer.ResourceServerAuthorizeRequestMatcherCustomizer;
 import com.seatliberator.seatliberator.bootstrap.security.customizer.ResourceServerHttpSecurityCustomizer;
 import com.seatliberator.seatliberator.bootstrap.security.customizer.ResourceServerOAuth2Customizer;
-import com.seatliberator.seatliberator.identity.client.actor.ThreadLocalActorContextHolder;
-import com.seatliberator.seatliberator.identity.client.role.NamespaceRoleCapabilitiesRegistry;
 import com.seatliberator.seatliberator.identity.core.actor.ActorContextHolder;
-import com.seatliberator.seatliberator.identity.core.role.NamespaceRoleDeserializer;
-import com.seatliberator.seatliberator.kernel.CurrentApplicationNamespaceProvider;
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
@@ -20,8 +14,10 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.AnnotationAwareOrderComparator;
 import org.springframework.core.annotation.Order;
+import org.springframework.core.convert.converter.Converter;
+import org.springframework.security.authentication.AbstractAuthenticationToken;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.security.oauth2.jwt.JwtDecoder;
-import org.springframework.security.oauth2.jwt.NimbusJwtDecoder;
 import org.springframework.security.oauth2.server.resource.web.authentication.BearerTokenAuthenticationFilter;
 
 import java.util.ArrayList;
@@ -40,7 +36,7 @@ public class ResourceServerAuthorizeAutoConfiguration {
     @Order(Ordered.HIGHEST_PRECEDENCE)
     ResourceServerOAuth2Customizer defaultResourceServerOAuth2Customizer(
             JwtDecoder jwtDecoder,
-            JwtAuthenticationTokenConverter jwtAuthenticationTokenConverter
+            Converter<Jwt, AbstractAuthenticationToken> jwtAuthenticationTokenConverter
     ) {
         return oauth -> oauth.jwt(jwt -> jwt
                 .decoder(jwtDecoder)
@@ -56,33 +52,6 @@ public class ResourceServerAuthorizeAutoConfiguration {
         return auth -> {
             for (var permit : properties.permits()) auth.requestMatchers(permit).permitAll();
         };
-    }
-
-    @Bean
-    @ConditionalOnMissingBean(JwtDecoder.class)
-    JwtDecoder jwtDecoder(ResourceServerAuthorizeProperties properties) {
-        if (properties.jwkSetUri() == null) {
-            throw new IllegalArgumentException(
-                    "seatliberator.resource-server.security.jwk-set-uri must not be null when enabled=true."
-            );
-        }
-        return NimbusJwtDecoder.withJwkSetUri(properties.jwkSetUri().toString()).build();
-    }
-
-    @Bean
-    @ConditionalOnMissingBean
-    ActorContextJwtAuthenticationConverter actorContextJwtAuthenticationConverter(
-            NamespaceRoleDeserializer namespaceRoleDeserializer,
-            NamespaceRoleCapabilitiesRegistry namespaceRoleCapabilitiesRegistry,
-            CurrentApplicationNamespaceProvider currentApplicationNamespaceProvider
-    ) {
-        return new ActorContextJwtAuthenticationConverter(namespaceRoleDeserializer, namespaceRoleCapabilitiesRegistry, currentApplicationNamespaceProvider);
-    }
-
-    @Bean
-    @ConditionalOnMissingBean(ActorContextHolder.class)
-    ThreadLocalActorContextHolder threadLocalActorContextHolder() {
-        return new ThreadLocalActorContextHolder();
     }
 
     @Bean
