--- bootstrap/resource-application-starter/src/main/java/com/seatliberator/seatliberator/bootstrap/security/JwtAuthenticationTokenConverter.java
@@ -1,9 +1,0 @@
-package com.seatliberator.seatliberator.bootstrap.security;
-
-import org.springframework.core.convert.converter.Converter;
-import org.springframework.security.authentication.AbstractAuthenticationToken;
-import org.springframework.security.oauth2.jwt.Jwt;
-
-@FunctionalInterface
-public interface JwtAuthenticationTokenConverter extends Converter<Jwt, AbstractAuthenticationToken> {
-}
