+++ identity/identity-client/src/main/java/com/seatliberator/seatliberator/identity/client/autoconfigure/IdentityClientProperties.java
@@ -1,0 +1,15 @@
+package com.seatliberator.seatliberator.identity.client.autoconfigure;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.bind.DefaultValue;
+
+import java.net.URI;
+
+@ConfigurationProperties(prefix = "identity.client")
+public record IdentityClientProperties(
+        @DefaultValue("true")
+        boolean enabled,
+
+        URI jwkSetUri
+) {
+}
