+++ identity/identity-client/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,2 +1,2 @@
 com.seatliberator.seatliberator.identity.client.introspector.web.WebIntrospectionAutoConfiguration
-com.seatliberator.seatliberator.identity.client.role.IdentityClientNamespaceRoleAutoConfiguration
+com.seatliberator.seatliberator.identity.client.autoconfigure.IdentityClientAutoConfiguration
