+++ reservation/reservation-application/src/main/resources/application-local.yml
@@ -1,8 +1,6 @@
-seatliberator:
-  resource-server:
-    security:
-      authorize:
-        jwk-set-uri: ${LOCAL_JWKS_BASE_URL:http://localhost:8080}/.well-known/jwks.json
+identity:
+  client:
+    jwk-set-uri: ${LOCAL_JWKS_BASE_URL:http://localhost:8080}/.well-known/jwks.json
 
 app:
   seed:
